Many changes - needs breakdown change by change
This branch seems to be recording experiments made
using the pwm function to drive an analog panel
meter.

Meter is centered on approx. 62F, and should have
enough range (in this very edit) for all common
conditions.  Reads about 37/50 (2/3 scale deflection)
at 73F and bottoms out in the upper 40's or low 50's.

Set point is 52F. A set point of 48F is to be taken
as difficult to achieve (true or untrue, it's an
assumption here).

On branch testing-aa-c
	modified:   tmp36-cc.ino